+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,31 @@
 # ui
 
+## 1.5.15-next.8
+
+### Patch Changes
+
+- 06f9f35: - feat(interrupt): add copilotkit interrupt as messages with copilotkit interrupt convenience fn
+  - chore(deps): update dependencies for demos
+  - chore(interrupt-as-message): add e2e test for interrupt as message
+- Updated dependencies [06f9f35]
+  - @copilotkit/runtime-client-gql@1.5.15-next.8
+  - @copilotkit/shared@1.5.15-next.8
+
+## 1.5.15-next.7
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.5.15-next.7
+- @copilotkit/shared@1.5.15-next.7
+
+## 1.5.15-next.6
+
+### Patch Changes
+
+- c9ae305: - perf: prevent redundant API calls
+  - @copilotkit/runtime-client-gql@1.5.15-next.6
+  - @copilotkit/shared@1.5.15-next.6
+
 ## 1.5.15-next.5
 
 ### Patch Changes