+++ CopilotKit/packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @copilotkit/sdk-js
 
+## 1.5.15-next.8
+
+### Patch Changes
+
+- 06f9f35: - feat(interrupt): add copilotkit interrupt as messages with copilotkit interrupt convenience fn
+  - chore(deps): update dependencies for demos
+  - chore(interrupt-as-message): add e2e test for interrupt as message
+  - @copilotkit/shared@1.5.15-next.8
+
+## 1.5.15-next.7
+
+### Patch Changes
+
+- @copilotkit/shared@1.5.15-next.7
+
+## 1.5.15-next.6
+
+### Patch Changes
+
+- @copilotkit/shared@1.5.15-next.6
+
 ## 1.5.15-next.5
 
 ### Patch Changes