+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkit/shared
 
+## 1.5.15-next.8
+
+## 1.5.15-next.7
+
+## 1.5.15-next.6
+
 ## 1.5.15-next.5
 
 ## 1.5.15-next.4